MyWebsite/requirements.txt
Django
pyttsx3
psutil
pyautogui
win10toast
pywin32
python-dotenv
pytesseract
pillow
pyperclip
